refactor: :recycle: move copy function from `TheActor` to `HomeView`
